Improve button layout and function
Problem still: possible to enter multiple decimal points. I need to figure out where to put the logic to note that the decimal point has already been entered. This has to happen after the display number is updated which is done by a component which returns from a switch case.